+++ src/main/java/dao/StadiumDao.java
@@ -1,13 +1,11 @@
 package dao;
 
-import mapper.StadiumMapper;
 import domain.Stadium;
-import org.apache.ibatis.session.SqlSession;
-import org.apache.ibatis.session.SqlSessionFactory;
+import mapper.StadiumMapper;
 
 import java.util.List;
 
-public class StadiumDao {
+public class StadiumDao extends AbstractMybatisDao {
 
     private static final StadiumDao INSTANCE = new StadiumDao();
 
@@ -15,14 +13,6 @@ public class StadiumDao {
     }
 
     public static StadiumDao getInstance() { return INSTANCE; }
-
-    private SqlSessionFactory sqlSessionFactory;
-    private SqlSession session;
-
-    public void setSqlSessionFactory(SqlSessionFactory sqlSessionFactory) {
-        this.sqlSessionFactory = sqlSessionFactory;
-        session = sqlSessionFactory.openSession();
-    }
 
     public void insert(Stadium stadium, boolean newSession) {
         if (newSession) {
@@ -36,14 +26,5 @@ public class StadiumDao {
     public List<Stadium> selectAll() {
         StadiumMapper mapper = session.getMapper(StadiumMapper.class);
         return mapper.selectAll();
-    }
-
-    public void rollback() {
-        session.rollback();
-    }
-
-    private void manageNewSession() {
-        session.close();
-        session = sqlSessionFactory.openSession();
     }
 }
